remove old story
