+++ stories/index.js
@@ -62,7 +62,7 @@ require('../src/FloatingNotification/docs/index.story');
 require('../src/FontUpgrade/docs/index.story');
 require('../src/FormField/docs/index.story');
 require('../src/GenericModalLayout/docs/index.story');
-require('../src/GoogleAddressInput/docs/indexx.story');
+require('../src/GoogleAddressInput/docs/index.story');
 require('../src/GooglePreview/docs/index.story');
 require('../src/Grid/docs');
 require('../src/Heading/docs/index.story');
